+++ podcasts/SleepTimerViewController.swift
@@ -2,6 +2,9 @@ import PocketCastsUtils
 import UIKit
 
 class SleepTimerViewController: SimpleNotificationsViewController {
+
+    @IBOutlet var settingsBtn: UIButton!
+
     @IBOutlet var plusFiveBtn: UIButton! {
         didSet {
             plusFiveBtn.setTitle(L10n.sleepTimerAdd5Mins, for: .normal)
@@ -210,6 +213,8 @@ class SleepTimerViewController: SimpleNotificationsViewController {
 
         customTimeStepper.tintColor = ThemeColor.playerContrast01()
         customEpisodeStepper.tintColor = ThemeColor.playerContrast01()
+
+        settingsBtn.tintColor = ThemeColor.playerContrast02()
     }
 
     private func updateSleepRemainingTime() {
@@ -271,6 +276,11 @@ class SleepTimerViewController: SimpleNotificationsViewController {
     }
 
     // MARK: - Sleep Timer Actions
+
+    @IBAction func settingsTapped(_ sender: Any) {
+        Analytics.track(.playerSleepTimerSettingsTapped)
+        NavigationManager.sharedManager.navigateTo(NavigationManager.settingsGeneralKey, data: [NavigationManager.settingsGeneralRowKey: GeneralSettingsViewController.TableRow.autoRestartSleepTimer])
+    }
 
     @IBAction func fiveMinutesTapped(_ sender: Any) {
         PlaybackManager.shared.setSleepTimerInterval(5.minutes)